Simplify random::gen_seed

+++ model/util/random.h
@@ -142,9 +142,9 @@ struct RNG {
     //@}
     
     uint64_t gen_seed() {
-        uint32_t low = m_rng();
-        uint32_t high = m_rng();
-        return ((uint64_t)high << 32) | (uint64_t)low;
+        uint64_t low = m_rng();
+        uint64_t high = m_rng();
+        return (high << 32) | low;
     }
     
     ///@brief Random number distributions
